[feat] update n_to_m to properly deal with masks
I thought I was masking the objects properly previously, but it turns
out that masking does not work well with the operations performed by
scattering (particularly, Complex values and the fft).  As an
alternative, I am doing inpainting of where the masked regions are.

Note that I have left some commented code in to do 'reconstruction',
though it did not work well. I will remove it in a subsequent CL.